* Updates README, HISTORY, LICENSE * Refactors protocol namespace * Moves StepHelper to step namespace * Updates unit tests
